less strict mode
Sometimes we need to change some minor thing in the source text or reference translation
(e.g. normalize white space).
It is pity to not be able to compare translations before and after such change.
On the other hand, we want to warn if there are some differences
because it means that the comparison may be unfair.